% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcol.R
\name{grid_spec}
\alias{grid_spec}
\alias{grid_spec}
\alias{grid_spec.default}
\alias{grid_spec}
\alias{grid_spec.BasicRaster}
\alias{grid_spec}
\alias{grid_spec.gridcol}
\alias{grid_spec}
\alias{grid_spec.gridspec}
\title{Title}
\usage{
grid_spec(x, ...)

\method{grid_spec}{default}(x, ...)

\method{grid_spec}{BasicRaster}(x, ...)

\method{grid_spec}{gridcol}(x, ...)

\method{grid_spec}{gridspec}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(raster)
grid_spec(raster(volcano))
grid_spec(grid_spec(raster(volcano)))
grid_spec(grid_spec(volcano))
gridcol(1, raster(volcano))
}
