% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcol.R, R/tibble.R
\name{gridcol}
\alias{gridcol}
\alias{gridcol}
\alias{gridcol.numeric}
\alias{gridcol}
\alias{gridcol.integer}
\alias{gridcol}
\alias{gc_tibble}
\alias{gridcol}
\alias{gc_tibble.BasicRaster}
\title{Title}
\usage{
gridcol(x, grid, ...)

\method{gridcol}{numeric}(x, grid, ...)

\method{gridcol}{integer}(x, grid, ...)

gc_tibble(x, ...)

\method{gc_tibble}{BasicRaster}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(raster)
gridcol(1, raster(volcano))
gridcol(1, volcano)
gridcol(seq_len(prod(dim(volcano))), volcano)
}
