% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcol.R
\name{[.gridcol}
\alias{[.gridcol}
\alias{[.gridcol}
\alias{[[.gridcol}
\title{gridcol methods}
\usage{
\method{[}{gridcol}(x, i)

\method{[[}{gridcol}(x, i)
}
\arguments{
\item{i}{}
}
\description{
gridcol methods
}
