% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcol.R
\name{x_coord}
\alias{x_coord}
\alias{x_coord}
\alias{x_coord.gridcol}
\alias{x_coord}
\alias{y_coord}
\alias{x_coord}
\alias{y_coord.gridcol}
\title{X coordinate}
\usage{
x_coord(x, ...)

\method{x_coord}{gridcol}(x, ...)

y_coord(x, ...)

\method{y_coord}{gridcol}(x, ...)
}
\arguments{
\item{x}{gridcol}

\item{...}{arguments for methods}
}
\description{
X coordinate
}
\examples{
x_coord(gridcol(1:10, raster(volcano)))
y_coord(gridcol(1:10, raster(volcano)))
}
