% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcol.R
\name{grid_spec}
\alias{grid_spec}
\alias{grid_spec}
\alias{grid_spec.default}
\alias{grid_spec}
\alias{grid_spec.BasicRaster}
\alias{grid_spec}
\alias{grid_spec.gridcol}
\title{Title}
\usage{
grid_spec(x, ...)

\method{grid_spec}{default}(x, ...)

\method{grid_spec}{BasicRaster}(x, ...)

\method{grid_spec}{gridcol}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
