% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridcol.R
\name{gridcol}
\alias{gridcol}
\alias{gridcol}
\alias{gridcol.numeric}
\alias{gridcol}
\alias{gridcol.integer}
\title{Title}
\usage{
gridcol(x, grid, ...)

\method{gridcol}{numeric}(x, grid, ...)

\method{gridcol}{integer}(x, grid, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
